Long으로 다바꿈

+++ src/main/java/bokathon4th/core/domain/InitialGame.java
@@ -11,7 +11,7 @@ public class InitialGame {
 
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Integer id;  // 글자 ID
+    private Long id;  // 글자 ID
 
     @Column(nullable = false)
     private String twoInitialWords;  // 두 글자 초성 (예: "ㅅㄱ")
